Refactor lobby components to use props instead of lobbyData; add LobbyDashboard for better structure
